feat(core): add FluxTableColumn.get(row)
